requirements.txt
# Copyright (c) 2016 The University of Manchester
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with this program.  If not, see <https://www.gnu.org/licenses/>.

SpiNNUtilities == 1!6.0.1
SpiNNMachine == 1!6.0.1
SpiNNMan == 1!6.0.1
SpiNNaker_PACMAN == 1!6.0.1
SpiNNaker_DataSpecification == 1!6.0.1
SpiNNFrontEndCommon == 1!6.0.1
SpiNNakerGraphFrontEnd == 1!6.0.1
pathos
